Move analyze function into base of causing package
